chore: update CI workflow
More updates from https://github.com/videojs/http-streaming/pull/989.
Makes it actually ready for usage now.

As the description says, it requires videojs-generate-karma-config@7 and
newer, karma@5, and a number in the nvmrc file.

We read the nvmrc to know the version to use in the CI action and it can
only use a number directly. Having a single source of truth is better
than having a signal that we're on an lts version.